NOT calling CloseHandle() on hData
This reverts commit f07cc708cd77d90bc4865c4bbe51f88df9df9ccc.

It was actually correct _without_ the call to CloseHandle. Attempts
to close the handle provided by GetClipboardData() cause an invalid
memory access exception to be thrown.

It seems that this is a "fake" handle, it references data on the
clipboard that is opened by the call to OpenClipboard, but it is
not a regular handle that references anything the kernel is explictly
tracking. The handle is (presumably!) automatically invalidated
when CloseClipboard() is called.