Allow Ctrl-C to interrupt `LineEnding::AsIs` as well
As unmodified line endings basically resulted in a single system call,
they were not interrupted by a call to `ExitProcess()` arriving after
the call to `WriteFile()` had kicked in. By forcing some chunking or
buffering even if no modifications will be made to the input, we can
abort without that delay.